revert

+++ src/main/java/com/soarclient/utils/mouse/MouseUtils.java
@@ -2,7 +2,7 @@ package com.soarclient.utils.mouse;
 
 public class MouseUtils {
 
-	public static boolean isInside(int mouseX, int mouseY, double x, double y, double width, double height) {
+	public static boolean isInside(double mouseX, double mouseY, double x, double y, double width, double height) {
 		return (mouseX > x && mouseX < (x + width)) && (mouseY > y && mouseY < (y + height));
 	}
 }
